Bugfixes and fourier transform application
Fixed signal variables being chosen as 'random' events to view and probe, as it caused a lot of issues

Added in FFT functionality, and an additive FT event component, that allows for viewing of general trends in the noise. Played around with real and imaginary components of dfthist, but decided against choosing either.