Created LCMS linear trendline removal
Applied the method from section 2.2 of https://inspirehep.net/literature/928989

To get more efficient and more accurate linear trendline removal method.

Currently there are 3 LCMS methods, and one least squares fitting method for linear trendline removal. LCMStest tries all 4 of these methods across y (10^5) data points and spits out a random event (to show the difference in results) as well as the time taken to execute.

LCMSlist is the fastest method from my testing, and requires a list of events (so an array of arrays) to function.

If you wish to remove linear trends for individual events, LCMS is the go to.